Move equation.py from fd/ to types/
