+++ devito/__init__.py
@@ -12,6 +12,7 @@ from devito.finite_differences import *  # noqa
 from devito.mpi import MPI  # noqa
 from devito.types import _SymbolCache, NODE, CELL, Buffer, SubDomain, SubDomainSet  # noqa
 from devito.types.dimension import *  # noqa
+from devito.types.equation import *  # noqa
 from devito.types.tensor import *  # noqa
 
 # Imports required to initialize Devito
