+++ devito/finite_differences/__init__.py
@@ -4,4 +4,3 @@ from .derivative import *  # noqa
 from .tools import generate_fd_shortcuts  # noqa
 from .coefficients import *  # noqa
 from .operators import *  # noqa
-from .equation import *  # noqa
