+++ devito/ir/equations/equation.py
@@ -4,9 +4,9 @@ from devito.ir.equations.algorithms import dimension_sort
 from devito.ir.support import (IterationSpace, DataSpace, Interval, IntervalGroup,
                                Stencil, detect_accesses, detect_oobs, detect_io,
                                build_intervals, build_iterators)
-from devito.finite_differences import Eq
 from devito.symbolics import FrozenExpr
 from devito.tools import Pickable, as_tuple
+from devito.types import Eq
 
 __all__ = ['LoweredEq', 'ClusterizedEq', 'DummyEq']
