+++ devito/operator/operator.py
@@ -7,7 +7,6 @@ from cached_property import cached_property
 import ctypes
 
 from devito.exceptions import InvalidOperator
-from devito.finite_differences import Eq
 from devito.logger import info, perf, warning, is_log_enabled_for
 from devito.ir.equations import LoweredEq
 from devito.ir.clusters import clusterize
@@ -20,7 +19,7 @@ from devito.symbolics import indexify
 from devito.targets import iet_lower
 from devito.tools import (DAG, Signer, ReducerMap, as_tuple, flatten, filter_ordered,
                           filter_sorted, split, timed_pass, timed_region)
-from devito.types import Dimension
+from devito.types import Dimension, Eq
 
 __all__ = ['Operator']
