+++ devito/types/__init__.py
@@ -1,14 +1,15 @@
 # Need only from within Devito
-from devito.types.basic import *  # noqa
+from .basic import *  # noqa
 
 # Needed both within and outside Devito
-from devito.types.dimension import *  # noqa
-from devito.types.utils import *  # noqa
-from devito.types.caching import _SymbolCache, CacheManager  # noqa
+from .dimension import *  # noqa
+from .utils import *  # noqa
+from .caching import _SymbolCache, CacheManager  # noqa
 
 # Needed only outside Devito
-from devito.types.constant import *  # noqa
-from devito.types.grid import *  # noqa
-from devito.types.dense import * # noqa
-from devito.types.sparse import *  # noqa
-from devito.types.tensor import *  # noqa
+from .equation import *  # noqa
+from .constant import *  # noqa
+from .grid import *  # noqa
+from .dense import * # noqa
+from .sparse import *  # noqa
+from .tensor import *  # noqa
