+++ devito/types/equation.py
