+++ devito/types/sparse.py
@@ -5,7 +5,7 @@ import sympy
 import numpy as np
 from cached_property import cached_property
 
-from devito.finite_differences import Differentiable, Eq, Inc, generate_fd_shortcuts
+from devito.finite_differences import Differentiable, generate_fd_shortcuts
 from devito.logger import warning
 from devito.mpi import MPI, SparseDistributor
 from devito.symbolics import INT, cast_mapper, indexify, retrieve_function_carriers
@@ -14,6 +14,7 @@ from devito.tools import (ReducerMap, flatten, prod, powerset,
 from devito.types.dense import DiscreteFunction, Function, SubFunction
 from devito.types.dimension import Dimension, ConditionalDimension, DefaultDimension
 from devito.types.basic import Symbol, Scalar
+from devito.types.equation import Eq, Inc
 
 __all__ = ['SparseFunction', 'SparseTimeFunction', 'PrecomputedSparseFunction',
            'PrecomputedSparseTimeFunction']
